Honour the "full" and "max_download" options.
Unless a local copy was explicitly indicated, access to a TarArchive
would repeatedly download the whole archive each time a single image
was requested. Now the TarArchive is configured on creation to
fully download or not. Typically if only a single frame is requested,
there is no need to download the whole archive, but if multiple
frames are requested, it is more efficient to get the whole
archive.

TODO: unpack the archive as well to avoid repeated searching
for specified frames in the packed archive.